refactor: move ExcelClinicalDataRepository to infrastructure layer
